requirements.txt
paho-mqtt
opencv-python
torch
vietocr
ultralytics
pymongo
flask
pillow